fix: frontend registration address is wrong, #4186
